[AP] output optimized

+++ SWIFTVehicle Renting System/Utility/Double+extension.swift
@@ -1,5 +1,5 @@
 import Foundation
-extension Float
+extension Double
 {
 func addDollarToBill()->String
     {
